pack AssetConfig into 2 storage slots, uint128 prices, gas optimizations
struct packing: 8 slots → 2 slots per asset
- bool configured + bool paused + uint16 cbThreshold + uint16 cumCap
  + uint40 lastTimestamp + uint40 anchorTimestamp + uint128 lastPrice = 32 bytes (slot 1)
- uint128 anchorPrice = 16 bytes (slot 2)

prices are uint128 (always positive, contract rejects zero).
getPrice returns int256 for Euler compatibility by casting on read.

additional optimizations:
- unchecked deviation math (safe: uint128 * 10000 fits uint256)
- unchecked loop increments
- cached block.timestamp as uint40
- cached maxPriceAge/anchorWindow as local vars
- reordered pause check before guards for short-circuit

gas results (vs original 8-slot layout):
  keeperPushPrices avg: 54,832 → 38,742 (-29%)
  getPrice avg:         7,080 → 5,117  (-28%)
  getAssetConfig:      15,792 → 5,598  (-65%)

90 tests pass.

+++ script/ConfigureDev.s.sol
@@ -103,12 +103,12 @@ contract ConfigureDev is Script {
 
         // convert fixed arrays to dynamic for the contract call
         address[] memory addrs = new address[](25);
-        uint256[] memory cbs = new uint256[](25);
-        uint256[] memory caps = new uint256[](25);
+        uint16[] memory cbs = new uint16[](25);
+        uint16[] memory caps = new uint16[](25);
         for (uint256 i = 0; i < 25; i++) {
             addrs[i] = tokens[i];
-            cbs[i] = cb[i];
-            caps[i] = cumCap[i];
+            cbs[i] = uint16(cb[i]);
+            caps[i] = uint16(cumCap[i]);
         }
 
         vm.startBroadcast(deployerKey);
